backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and ML dependencies
google-generativeai==0.3.2
langchain==0.0.335
langchain-community==0.0.10
transformers==4.35.2
torch==2.1.1
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2

# Data sources and trading
yfinance==0.2.28
ccxt==4.1.48
alpha-vantage==2.3.1
requests==2.31.0

# Database and caching  
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Visualization and analysis
matplotlib==3.8.2
plotly==5.17.0
ta==0.10.2
mplfinance==0.12.10b0

# WebSocket and real-time
python-socketio==5.10.0

# Security and validation
pydantic==2.5.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
croniter==1.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0